adds Pundit policies for VAT, stripe-rails, stripe-ruby-mock and invoice mailer

+++ app/controllers/vat_rates_controller.rb
@@ -1,19 +1,23 @@
 class VatRatesController < ApplicationController
 
   before_action :authenticate_user!
-
+  after_action :verify_authorized
+  after_action :verify_policy_scoped, except: [:new, :create]
+  
   def index
-    @vat_rates = VatRate.visible_to(current_user).order(name: :asc )
+    @vat_rates = policy_scope(VatRate).order(name: :asc )
+    authorize @vat_rates
     @vat_rates = @vat_rates.active unless params[:all]
-    # @vat_rates = current_account.vat_rates if current_account
   end
 
   def new
     @vat_rate = current_account.vat_rates.build(active: true)
+    authorize @vat_rate
   end
 
   def create
     @vat_rate = VatRate.new(vat_params)
+    authorize @vat_rate
     if @vat_rate.save
       redirect_to vat_rates_url, notice: "VAT rate successfully added"
     else
@@ -22,11 +26,13 @@ class VatRatesController < ApplicationController
   end
 
   def edit
-    @vat_rate = VatRate.visible_to(current_user).find(params[:id])
+    @vat_rate = policy_scope(VatRate).find(params[:id])
+    authorize @vat_rate
   end
 
   def update
-    @vat_rate = VatRate.visible_to(current_user).find(params[:id])
+    @vat_rate = policy_scope(VatRate).find(params[:id])
+    authorize @vat_rate
     if @vat_rate.update_attributes(vat_params)
       redirect_to vat_rates_url, notice: "VAT rate successfully updated"
     else
@@ -35,7 +41,8 @@ class VatRatesController < ApplicationController
   end
 
   def destroy
-    @vat_rate = VatRate.visible_to(current_user).find(params[:id])
+    @vat_rate = policy_scope(VatRate).find(params[:id])
+    authorize @vat_rate
     if @vat_rate.destroy
       flash[:success] = "VAT rate successfully deleted"
       redirect_to :vat_rates
@@ -49,7 +56,7 @@ class VatRatesController < ApplicationController
   private
 
   def vat_params
-    params.require(:vat_rate).permit(:account_id, :name, :rate, :active)
+    params.require(:vat_rate).permit(*policy(@vat_rate || VatRate).permitted_attributes)
   end
 
 end
